+++ test/bench/bench_mlkem.c
@@ -7,11 +7,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "../mlkem/src/common.h"
-#include "../mlkem/src/randombytes.h"
+#include "../../mlkem/src/common.h"
+#include "../../mlkem/src/randombytes.h"
 #include "hal.h"
 
-#include "../mlkem/mlkem_native.h"
+#include "../../mlkem/mlkem_native.h"
 
 #define NWARMUP 50
 #define NITERATIONS 300
